consumer/requirements.txt
confluent-kafka==1.8.2
cassandra-driver==3.25.0